+++ idunn/blocks/__init__.py
@@ -3,7 +3,7 @@ from .phone import PhoneBlock
 from .information import InformationBlock
 from .website import WebSiteBlock
 from .contact import ContactBlock
-from .wikipedia import WikipediaBlock
+from .wikipedia import WikipediaBlock, WikiUndefinedException, GET_WIKI_INFO, WikipediaCache
 from .images import ImagesBlock
 from .services_and_information import ServicesAndInformationBlock, AccessibilityBlock, InternetAccessBlock, BreweryBlock
